+++ plugins/wp-chargify/wp-chargify.php
@@ -30,10 +30,10 @@ require_once( 'includes/admin/settings.php' );
 require_once( 'includes/admin/helpers.php' );
 require_once( 'includes/chargify/endpoints/product-families.php' );
 require_once( 'includes/chargify/endpoints/subscriptions.php' );
+require_once('includes/chargify/endpoints/webhooks.php');
 require_once( 'includes/chargify/customers/namespace.php' );
 require_once( 'includes/chargify/renewal/namespace.php' );
 require_once( 'includes/chargify/subscription/namespace.php' );
-require_once( 'includes/chargify/webhooks/namespace.php' );
 require_once( 'includes/endpoints/base.php' );
 require_once( 'includes/forms/account-details.php' );
 require_once( 'includes/forms/billing-details.php' );
